cahnge uint64

+++ contracts/FullCheckpoint.sol
@@ -32,8 +32,8 @@ contract FullCheckpoint {
     }
 
     struct Range {
-        uint256 lastFinalizedNumber;
-        uint256 latestFinalizedNumber;
+        uint64 lastFinalizedNumber;
+        uint64 latestFinalizedNumber;
     }
 
     Range[] public finalizedRanges;
@@ -125,7 +125,7 @@ contract FullCheckpoint {
         finalizedRanges.push(
             Range({
                 lastFinalizedNumber: 0,
-                latestFinalizedNumber: uint256(gapBlock.number)
+                latestFinalizedNumber: uint64(uint256(gapBlock.number))
             })
         );
 
@@ -309,11 +309,11 @@ contract FullCheckpoint {
     function setCommittedStatus(bytes32 committedBlock) internal {
         finalizedRanges.push(
             Range({
-                lastFinalizedNumber: uint256(
-                    uint64(headerTree[latestFinalizedBlock].mix >> 128)
+                lastFinalizedNumber: uint64(
+                    headerTree[latestFinalizedBlock].mix >> 128
                 ),
-                latestFinalizedNumber: uint256(
-                    uint64(headerTree[committedBlock].mix >> 128)
+                latestFinalizedNumber: uint64(
+                    headerTree[committedBlock].mix >> 128
                 )
             })
         );
@@ -438,7 +438,7 @@ contract FullCheckpoint {
      */
     function getFinalizedInfo(
         bytes32 blockHash
-    ) public view returns (int256 finalizedNumber, bool isFinalized) {
+    ) public view returns (int64 finalizedNumber, bool isFinalized) {
         uint256 blockNumber = uint256(uint64(headerTree[blockHash].mix >> 128));
         if (blockNumber != 0) {
             for (uint256 i = 0; i < finalizedRanges.length; i++) {
@@ -447,7 +447,7 @@ contract FullCheckpoint {
                     blockNumber <= finalizedRanges[i].latestFinalizedNumber
                 ) {
                     return (
-                        int256(finalizedRanges[i].latestFinalizedNumber),
+                        int64(finalizedRanges[i].latestFinalizedNumber),
                         true
                     );
                 }
